fix: use i64 for stream stats

+++ src/common/meta/organization.rs
@@ -64,8 +64,8 @@ pub struct OrgSummary {
 #[derive(Serialize, Deserialize, ToSchema)]
 pub struct StreamSummary {
     pub num_streams: i64,
-    pub total_storage_size: f64,
-    pub total_compressed_size: f64,
+    pub total_storage_size: i64,
+    pub total_compressed_size: i64,
 }
 
 /// A container for passcodes and rumtokens
