+++ src/infra/src/cache/stats.rs
@@ -78,8 +78,8 @@ pub fn incr_stream_stats(key: &str, val: &FileMeta) -> Result<(), anyhow::Error>
     }
     stats.doc_num += val.records;
     stats.file_num += 1;
-    stats.storage_size += val.original_size as f64;
-    stats.compressed_size += val.compressed_size as f64;
+    stats.storage_size += val.original_size;
+    stats.compressed_size += val.compressed_size;
 
     Ok(())
 }
@@ -112,8 +112,8 @@ mod tests {
             doc_time_max: 1667978845374,
             doc_num: 5000,
             file_num: 1,
-            storage_size: 200.00,
-            compressed_size: 3.00,
+            storage_size: 200,
+            compressed_size: 3,
         };
 
         set_stream_stats("nexus", "default", StreamType::Logs, val);
